+++ rec/tests/test_PopularityRecommender.py
@@ -1,15 +1,15 @@
 from rec.models import PopularityRecommender
 import numpy as np
 import pytest
-import test_utils
+import test_helpers
 
 
 class TestPopularityRecommender:
     def test_default(self):
         c = PopularityRecommender()
-        test_utils.assert_correct_num_users(c.num_users, c, c.user_profiles.shape[0])
-        test_utils.assert_correct_num_items(c.num_items, c, c.item_attributes.shape[1])
-        test_utils.assert_not_none(c.predicted_scores)
+        test_helpers.assert_correct_num_users(c.num_users, c, c.user_profiles.shape[0])
+        test_helpers.assert_correct_num_items(c.num_items, c, c.item_attributes.shape[1])
+        test_helpers.assert_not_none(c.predicted_scores)
 
     def test_arguments(self, items=None, users=None):
         if items is None:
@@ -19,9 +19,9 @@ class TestPopularityRecommender:
 
         # init with given arguments
         c = PopularityRecommender(num_users=users, num_items=items)
-        test_utils.assert_correct_num_users(users, c, c.user_profiles.shape[0])
-        test_utils.assert_correct_num_items(items, c, c.item_attributes.shape[1])
-        test_utils.assert_not_none(c.predicted_scores)
+        test_helpers.assert_correct_num_users(users, c, c.user_profiles.shape[0])
+        test_helpers.assert_correct_num_items(items, c, c.item_attributes.shape[1])
+        test_helpers.assert_not_none(c.predicted_scores)
 
     def test_partial_arguments(self, items=None, users=None):
         # init with partially given arguments
@@ -31,19 +31,19 @@ class TestPopularityRecommender:
             users = np.random.randint(1, 100)
 
         c = PopularityRecommender(num_users=users)
-        test_utils.assert_correct_num_users(users, c, c.user_profiles.shape[0])
-        test_utils.assert_correct_num_items(c.num_items, c, c.item_attributes.shape[1])
-        test_utils.assert_not_none(c.predicted_scores)
+        test_helpers.assert_correct_num_users(users, c, c.user_profiles.shape[0])
+        test_helpers.assert_correct_num_items(c.num_items, c, c.item_attributes.shape[1])
+        test_helpers.assert_not_none(c.predicted_scores)
 
         c = PopularityRecommender(num_items=items)
-        test_utils.assert_correct_num_users(c.num_users, c, c.user_profiles.shape[0])
-        test_utils.assert_correct_num_items(items, c, c.item_attributes.shape[1])
-        test_utils.assert_not_none(c.predicted_scores)
+        test_helpers.assert_correct_num_users(c.num_users, c, c.user_profiles.shape[0])
+        test_helpers.assert_correct_num_items(items, c, c.item_attributes.shape[1])
+        test_helpers.assert_not_none(c.predicted_scores)
 
         c = PopularityRecommender(num_users=users, num_items=items)
-        test_utils.assert_correct_num_users(users, c, c.user_profiles.shape[0])
-        test_utils.assert_correct_num_items(items, c, c.item_attributes.shape[1])
-        test_utils.assert_not_none(c.predicted_scores)
+        test_helpers.assert_correct_num_users(users, c, c.user_profiles.shape[0])
+        test_helpers.assert_correct_num_items(items, c, c.item_attributes.shape[1])
+        test_helpers.assert_not_none(c.predicted_scores)
 
     def test_representations(self, item_repr=None, user_repr=None):
         if item_repr is None:
@@ -54,33 +54,33 @@ class TestPopularityRecommender:
             user_repr = np.random.randint(10, size=(users, 1))
 
         c = PopularityRecommender(item_representation=item_repr)
-        test_utils.assert_correct_num_users(c.num_users, c, c.user_profiles.shape[0])
-        test_utils.assert_correct_num_items(
+        test_helpers.assert_correct_num_users(c.num_users, c, c.user_profiles.shape[0])
+        test_helpers.assert_correct_num_items(
             item_repr.shape[1], c, c.item_attributes.shape[1]
         )
-        test_utils.assert_equal_arrays(item_repr, c.item_attributes)
-        test_utils.assert_not_none(c.predicted_scores)
+        test_helpers.assert_equal_arrays(item_repr, c.item_attributes)
+        test_helpers.assert_not_none(c.predicted_scores)
 
         c = PopularityRecommender(user_representation=user_repr)
-        test_utils.assert_correct_num_users(
+        test_helpers.assert_correct_num_users(
             user_repr.shape[0], c, c.user_profiles.shape[0]
         )
-        test_utils.assert_correct_num_items(c.num_items, c, c.item_attributes.shape[1])
-        test_utils.assert_equal_arrays(user_repr, c.user_profiles)
-        test_utils.assert_not_none(c.predicted_scores)
+        test_helpers.assert_correct_num_items(c.num_items, c, c.item_attributes.shape[1])
+        test_helpers.assert_equal_arrays(user_repr, c.user_profiles)
+        test_helpers.assert_not_none(c.predicted_scores)
 
         c = PopularityRecommender(
             user_representation=user_repr, item_representation=item_repr
         )
-        test_utils.assert_correct_num_users(
+        test_helpers.assert_correct_num_users(
             user_repr.shape[0], c, c.user_profiles.shape[0]
         )
-        test_utils.assert_correct_num_items(
+        test_helpers.assert_correct_num_items(
             item_repr.shape[1], c, c.item_attributes.shape[1]
         )
-        test_utils.assert_equal_arrays(user_repr, c.user_profiles)
-        test_utils.assert_equal_arrays(item_repr, c.item_attributes)
-        test_utils.assert_not_none(c.predicted_scores)
+        test_helpers.assert_equal_arrays(user_repr, c.user_profiles)
+        test_helpers.assert_equal_arrays(item_repr, c.item_attributes)
+        test_helpers.assert_not_none(c.predicted_scores)
 
     def test_wrong_representation(
         self, user_repr=None, item_repr=None, bad_user_repr=None, bad_item_repr=None
@@ -119,26 +119,26 @@ class TestPopularityRecommender:
         c = PopularityRecommender(verbose=False, num_items_per_iter=num_items_per_iter)
         assert num_items_per_iter == c.num_items_per_iter
         # also check other params
-        test_utils.assert_correct_num_users(c.num_users, c, c.user_profiles.shape[0])
-        test_utils.assert_correct_num_items(c.num_items, c, c.item_attributes.shape[1])
-        test_utils.assert_not_none(c.predicted_scores)
+        test_helpers.assert_correct_num_users(c.num_users, c, c.user_profiles.shape[0])
+        test_helpers.assert_correct_num_items(c.num_items, c, c.item_attributes.shape[1])
+        test_helpers.assert_not_none(c.predicted_scores)
 
     def test_seeding(self, seed=None, items=None, users=None):
         if seed is None:
             seed = np.random.randint(100000)
         s1 = PopularityRecommender(seed=seed)
         s2 = PopularityRecommender(seed=seed)
-        test_utils.assert_equal_arrays(s1.item_attributes, s2.item_attributes)
-        test_utils.assert_equal_arrays(s1.user_profiles, s2.user_profiles)
+        test_helpers.assert_equal_arrays(s1.item_attributes, s2.item_attributes)
+        test_helpers.assert_equal_arrays(s1.user_profiles, s2.user_profiles)
         s1.run(timesteps=5)
         s2.run(timesteps=5)
         # check that measurements are the same
         meas1 = s1.get_measurements()
         meas2 = s2.get_measurements()
-        test_utils.assert_equal_measurements(meas1, meas2)
+        test_helpers.assert_equal_measurements(meas1, meas2)
         systate1 = s1.get_system_state()
         systate2 = s2.get_system_state()
-        test_utils.assert_equal_system_state(systate1, systate2)
+        test_helpers.assert_equal_system_state(systate1, systate2)
 
         if items is None:
             items = np.random.randint(10, 1000)
@@ -146,14 +146,14 @@ class TestPopularityRecommender:
             users = np.random.randint(10, 100)
         s1 = PopularityRecommender(seed=seed, num_users=users, num_items=items)
         s2 = PopularityRecommender(seed=seed, num_users=users, num_items=items)
-        test_utils.assert_equal_arrays(s1.item_attributes, s2.item_attributes)
-        test_utils.assert_equal_arrays(s1.user_profiles, s2.user_profiles)
+        test_helpers.assert_equal_arrays(s1.item_attributes, s2.item_attributes)
+        test_helpers.assert_equal_arrays(s1.user_profiles, s2.user_profiles)
         s1.run(timesteps=5)
         s2.run(timesteps=5)
         # check that measurements are the same
         meas1 = s1.get_measurements()
         meas2 = s2.get_measurements()
-        test_utils.assert_equal_measurements(meas1, meas2)
+        test_helpers.assert_equal_measurements(meas1, meas2)
         systate1 = s1.get_system_state()
         systate2 = s2.get_system_state()
-        test_utils.assert_equal_system_state(systate1, systate2)
+        test_helpers.assert_equal_system_state(systate1, systate2)
